% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vy_functions.R
\name{cdfun}
\alias{cdfun}
\title{Person fit index for jags markov model, currently breaks with bs items}
\usage{
cdfun(ydat, tord, p = 10, q = 10, static = FALSE)
}
\arguments{
\item{ydat}{y data in item order, anx1, anx2, ... sed10}

\item{tord}{the tord of the ymat}

\item{p}{the number of items per factor}

\item{q}{the number of factors}
}
\value{
the N by k pf matrix
}
\description{
Person fit index for jags markov model, currently breaks with bs items
}
