% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cond}
\alias{cond}
\title{treatment conditions of participants}
\format{
An object of class \code{numeric} of length 1002.
}
\usage{
cond
}
\description{
Short description:
}
\details{
cond1 = business as usual

cond2 = algorithm

cond3 = finish fast

cond4 = trick our algorithm but be fast


Full instructions:

cond2:
Note that this survey employs an advanced statistical technique that detects inattentive behavior. Please pay attention to the questions asked and respond carefully.

cond3:
Please respond to all subsequent items without effort but pretend that you want your inattention in filling out this survey to remain undetected.

cond4:
Please respond to all subsequent items without effort with and with no risk of penalty: in fact, we request that you do so.
}
\keyword{datasets}
