% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.instr}
\alias{df.instr}
\title{confidence if instruction were followed based on follow-up question}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 502 rows and 4 columns.
}
\usage{
df.instr
}
\description{
Confidence is ranked from 1 = high, 2 = no info,
3 = indication that instructions were not followed including that
instructions meant thinking less about answer or answering just
faster/more intuitively
}
\keyword{datasets}
