% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull data.R
\name{gen_df}
\alias{gen_df}
\title{generate data.frame with survey data}
\usage{
gen_df(dat.list = NULL)
}
\arguments{
\item{dat.list}{list of data, can be generated with pull_data, if no dat.list ist provided, the full data from the package is used}
}
\value{
data.frame
}
\description{
generate data.frame with survey data
}
