% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vy_functions.R
\name{getCprobs}
\alias{getCprobs}
\title{Extracts computes and formats Jags model to plot format data}
\usage{
getCprobs(RDSjags, cd = TRUE)
}
\arguments{
\item{RDSjags}{JAGS output model as list with inputs}

\item{cd}{Defaults to TRUE, was the CD function included in the markov model}
}
\value{
Returns the plotting DF of a post-proceed JAGS model
}
\description{
Extracts computes and formats Jags model to plot format data
}
