% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{JagsReport}
\alias{JagsReport}
\title{Generates post-proc report for jags model results}
\usage{
JagsReport(path)
}
\arguments{
\item{path}{location of jags output object as defined}
}
\value{
Returns pdf document report
}
\description{
Generates post-proc report for jags model results
}
