% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{JagsReportMultiCon}
\alias{JagsReportMultiCon}
\title{Generates post-proc report for jags model results with multiple conditions}
\usage{
JagsReportMultiCon(path)
}
\arguments{
\item{path}{location of jags output object as defined}
}
\value{
Returns pdf document report
}
\description{
Generates post-proc report for jags model results with multiple conditions
}
