% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vy_functions.R
\name{makevy2}
\alias{makevy2}
\title{Make moving variance function correct}
\usage{
makevy2(ymat, iord, tord, startLag = 4)
}
\arguments{
\item{ymat}{y data in item order, anx1, anx2, ... sed10}

\item{iord}{the iord of the ymat}

\item{tord}{the tord of the ymat}

\item{startLag}{how many items to skip before starting moving variance calculation, defaults to 4, should probably be higher.}
}
\value{
Returns the moving variance matrix in item presentation order.
}
\description{
Make moving variance function correct
}
