% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc analysis.R
\name{plot_states_shifted}
\alias{plot_states_shifted}
\title{Plot shifted states plot with treatment timing at 0}
\usage{
plot_states_shifted(df, summary = TRUE)
}
\arguments{
\item{df}{data.frame that contains
"treatment.timing",
"id" (person identifier),
".value" (value to plot, normally posterior mean),
"t" (item timing)}

\item{summary}{should summary stats be plotted}
}
\description{
Plot shifted states plot with treatment timing at 0
}
