% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc analysis.R
\name{plot_worst_trace}
\alias{plot_worst_trace}
\title{Plots traceplots of worst mixing parameters}
\usage{
plot_worst_trace(
  mcmc = NULL,
  summary = NULL,
  names = NULL,
  regex = NULL,
  n = 12
)
}
\arguments{
\item{mcmc}{mcmc output}

\item{summary}{summary output with Rhat column}

\item{names}{names of variables to consider, if NULL considers all}

\item{regex}{regex expression to find names}

\item{n}{number of parameters to plot}
}
\value{
ggplot
}
\description{
Plots traceplots of worst mixing parameters
}
