% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull data.R
\name{pull_data}
\alias{pull_data}
\title{Pull package data of online survey}
\usage{
pull_data(
  choice.cond = 4,
  bs = FALSE,
  n = NULL,
  standardize = TRUE,
  treatment.timing.min = -Inf,
  treatment.timing.max = Inf,
  choice.id = NULL
)
}
\arguments{
\item{choice.cond}{treatment condition to include (standard only 4)}

\item{bs}{use bs items (standard FALSE)}

\item{n}{number of persons to be sampled}

\item{standardize}{standardize item values (standard TRUE)}

\item{treatment.timing.min}{include only}

\item{treatment.timing.max}{}
}
\value{
list with data entries
}
\description{
Returns relevant data in list. Contains the following elements:
\itemize{
  \item{ymat[person,item] with item values}
  \item{cond[person] with treatment condition}
  \item{iord[person,time] with item index}
  \item{tord[person,item] with time shown of item}
  \item{treatment.timing[person] with timing of treatment}
  \item{cd[person,item] with timing of treatment}
  \item{instr[person] with response to instruction question}
}
}
\examples{
dat <- pull_data()
dim(dat$ymat)
all(colMeans(dat$ymat)<1e-10)
ls(dat)

dim(dat$cd)

### check order
# show first person
dat$raw_ord[1]
# check first item shown (t=1)
which(dat$tord[1,]==1)
dat$iord[1,1]
# check last item shown (t=100)
colnames(ymat)[dat$iord[1,100]]

#check place of first item (i=1)
dat$tord[1,1]

}
