% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vy_functions.R
\name{stfun}
\alias{stfun}
\title{Repeated string values for jags markov model, currently breaks with bs items}
\usage{
stfun(ydat, tord, p = 10, q = 10)
}
\arguments{
\item{ydat}{y data in item order, anx1, anx2, ... sed10}

\item{tord}{the tord of the ymat}

\item{p}{the number of items per factor}

\item{q}{the number of factors}
}
\value{
Returns the string function which provides the number of concurrent items which are in the middle category of 3 or 4
}
\description{
Repeated string values for jags markov model, currently breaks with bs items
}
