% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract mcmc.R
\name{subset_summary}
\alias{subset_summary}
\title{Subset mcmc.summary and add dimensions}
\usage{
subset_summary(summary, variable, dim = NULL, iden.column = "variable")
}
\arguments{
\item{summary}{mcmc.summary data.frame (but can be any df)}

\item{variable}{character to subset variable names in iden.column}

\item{dim}{vector with dimension names. optional.}

\item{iden.column}{name of column that contains variable names}
}
\description{
Can parse columns with entries like "variable[1,2]" into multiple columns containing the dimensions
}
